// NextJS/PortfolioApp/types/project.ts
export type Project = {
    _id: string;
    profile: string;
    institution: string;
    titlePT: string;
    titleEN: string;
    titleES: string;
    titleFR: string;
    detailsPT: string;
    detailsEN: string;
    detailsFR: string;
    detailsES: string;
    tecnologies:string;
    imgs:string[];
    url:string;
    active:boolean;
}